glibc-tools: Change the way to parellelize
This patch changes how to parallelize multiple glibc build.  The glibc
build contains a lot of choke points which serializes most of its
executions in both build and check mode:

  - Build is done each directory at time and parallel build opportunities
    only happens for multiple objects inside a directory.

  - Linking the libraries requires previous libraries.

  - nptl checks are explicit done in serial mode.

This patch add the '-p' option which expects either a number or a
number-number.  The first argument is the total number of build done
in parallel, where the optional second argument if the number of
jobs used for issue make (-j).  So a "-p 24:6" will issue 24 builds
in parallel using "make -j6".

This improves the ABI build on a 64 cores machines from 40 minutes
to 8 minutes.